key_manager/requirements.txt
flask
flask_sqlalchemy
itsdangerous
python-dateutil
PyYAML
db-sqlite3
